Clean up everthing
The old directory architectur is really ugly, tune everthing to fit a normal
Unix style.

With this change, we are abole to compile the book simply with the familiar:

$ make

Read it with 'make read', release it with 'make release'.

Now, do our configuration in configs/, read documents under doc/ and modify
tools in tools/, everything gets its right place.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>